+++ Modules/booking/Model/BkScheduling.php
@@ -32,15 +32,29 @@ class BkScheduling extends Model {
         $this->setColumnsInfo("resa_time_setting", "int(1)", 1);
         $this->setColumnsInfo("default_color_id", "int(11)", 1);
         $this->setColumnsInfo("id_rearea", "int(11)", 0);
+        $this->setColumnsInfo("force_packages", "tinyint", 0);
         $this->primaryKey = "id";
     }
 
     public function getDefault() {
-        return array("id" => 0, "is_monday" => 1, "is_tuesday" => 1,
-            "is_wednesday" => 1, "is_thursday" => 1, "is_friday" => 1,
-            "is_saturday" => 0, "is_sunday" => 0, "day_begin" => 8,
-            "day_end" => 18, "size_bloc_resa" => 3600, "booking_time_scale" => 1,
-            "resa_time_setting" => 1, "default_color_id" => 1, "id_rearea" => 0);
+        return array(
+            "id" => 0,
+            "is_monday" => 1,
+            "is_tuesday" => 1,
+            "is_wednesday" => 1,
+            "is_thursday" => 1,
+            "is_friday" => 1,
+            "is_saturday" => 0,
+            "is_sunday" => 0,
+            "day_begin" => 8,
+            "day_end" => 18,
+            "size_bloc_resa" => 3600,
+            "booking_time_scale" => 1,
+            "resa_time_setting" => 1,
+            "default_color_id" => 1,
+            "id_rearea" => 0,
+            "force_packages" => 0
+        );
     }
 
     
@@ -123,13 +137,13 @@ class BkScheduling extends Model {
      * @param string $name name of the SyColorCode
      * @param string $address address of the SyColorCode
      */
-    public function add($id_space, $id_rearea, $is_monday, $is_tuesday, $is_wednesday, $is_thursday, $is_friday, $is_saturday, $is_sunday, $day_begin, $day_end, $size_bloc_resa, $booking_time_scale, $resa_time_setting, $default_color_id) {
+    public function add($id_space, $id_rearea, $is_monday, $is_tuesday, $is_wednesday, $is_thursday, $is_friday, $is_saturday, $is_sunday, $day_begin, $day_end, $size_bloc_resa, $booking_time_scale, $resa_time_setting, $default_color_id,$force_packages=0) {
 
         $sql = "insert into bk_schedulings(is_monday, is_tuesday, "
                 . " is_wednesday, is_thursday, is_friday, is_saturday, is_sunday, day_begin,"
                 . " day_end, size_bloc_resa, booking_time_scale,"
-                . " resa_time_setting, default_color_id, id_space, id_rearea)"
-                . " values(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)";
+                . " resa_time_setting, default_color_id, id_space, id_rearea, force_packages)"
+                . " values(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)";
         $this->runRequest($sql, array($is_monday, $is_tuesday,
             $is_wednesday, $is_thursday, $is_friday, $is_saturday, $is_sunday, $day_begin,
             $day_end, $size_bloc_resa, $booking_time_scale,
@@ -144,15 +158,15 @@ class BkScheduling extends Model {
      * @param string $name New name of the SyColorCode
      * @param string $color New Address of the SyColorCode
      */
-    public function update2($id_space, $id_rearea, $is_monday, $is_tuesday, $is_wednesday, $is_thursday, $is_friday, $is_saturday, $is_sunday, $day_begin, $day_end, $size_bloc_resa, $booking_time_scale, $resa_time_setting, $default_color_id) {
+    public function update2($id_space, $id_rearea, $is_monday, $is_tuesday, $is_wednesday, $is_thursday, $is_friday, $is_saturday, $is_sunday, $day_begin, $day_end, $size_bloc_resa, $booking_time_scale, $resa_time_setting, $default_color_id, $force_packages=0) {
 
         $sql = "UPDATE bk_schedulings SET is_monday=?, is_tuesday=?, is_wednesday=?, is_thursday=?, is_friday=?, "
                 . "is_saturday=?, is_sunday=?, day_begin=?, day_end=?, size_bloc_resa=?, booking_time_scale=?, "
-                . "resa_time_setting=?, default_color_id=? "
+                . "resa_time_setting=?, default_color_id=?, force_packages=? "
                 . "WHERE id_rearea=? AND deleted=0 AND id_space=?";
         $this->runRequest($sql, array($is_monday, $is_tuesday, $is_wednesday, $is_thursday, $is_friday,
             $is_saturday, $is_sunday, $day_begin, $day_end, $size_bloc_resa,
-            $booking_time_scale, $resa_time_setting, $default_color_id, $id_rearea, $id_space));
+            $booking_time_scale, $resa_time_setting, $default_color_id, $force_packages, $id_rearea, $id_space));
     }
 
     protected function onToBool($on){
@@ -180,7 +194,9 @@ class BkScheduling extends Model {
         $size_bloc_resa,
         $booking_time_scale,
         $resa_time_setting,
-        $default_color_id) {
+        $default_color_id,
+        $force_packages
+    ) {
         $id = 0;
         if ($id = $this->existsByReArea($id_space, $id_rearea)) {
             $this->update2(
@@ -198,7 +214,9 @@ class BkScheduling extends Model {
                 $size_bloc_resa,
                 $booking_time_scale,
                 $resa_time_setting,
-                $default_color_id);
+                $default_color_id,
+                $force_packages
+            );
         } else {
             $id = $this->add(
                 $id_space,
@@ -215,7 +233,9 @@ class BkScheduling extends Model {
                 $size_bloc_resa,
                 $booking_time_scale,
                 $resa_time_setting,
-                $default_color_id);
+                $default_color_id,
+                $force_packages
+            );
         }
         return $id;
     }