+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * [users] if date_end_contract is empty, insert null in db
 * [core] code cleanup #554
 * [booking] fix package display on new reservation
+* [booking] add to scheduling option to force usage of packages to an area
 
 ## 2.3.2